(PC-41147)[API] refactor: separate movie routes in new folder
