+++ api/src/pcapi/routes/native/v1/__init__.py
@@ -13,6 +13,7 @@ def install_routes(app: Flask) -> None:
     from . import favorites
     from . import feedbacks
     from . import geography
+    from . import movies
     from . import offerers
     from . import offers
     from . import reaction
